// src/app/models/tests/test-info.ts
export enum TestType{
  GYROSCOPE = " GYROSCOPE_TEST",
  ACCELEROMETER = "ACCELEROMETER_TEST",
  FINGER_TAPPING = "FINGER_TAPPING",
  STATIC = "STATIC_TEST",
  VOICE = "VOICE_TEST",
  DEFAULT = ""
}

export enum TestIcon {
  FINGER_TAPPING = "fingerprint",
  GYROSCOPE = "watch",
  STATIC = "accessibility",
  ACCELEROMETER = "airline_seat_legroom_reduced",
  VOICE = "mic",
  DEFAULT = ""
}

export enum TestName {
  FINGER_TAPPING = "Stukanie palcami",
  GYROSCOPE = "Badanie drżeń",
  STATIC = "Test posturalny",
  ACCELEROMETER = "Stukanie nogą",
  VOICE = "Badanie głosu",

  DEFAULT = ""

}

export enum TestNameEng {
  FINGER_TAPPING = "Finger Tapping",
  GYROSCOPE = "Gyroscope Test",
  STATIC = "Static Test",
  ACCELEROMETER = "Accelerometer Test",
  VOICE = "Voice Test",

  DEFAULT = ""

}